chore: complete auth rename
