+++ .lintstagedrc
@@ -1,3 +1,3 @@
 {
-  "packages/server/**/*.+(js|ts)": ["npm run test:e2e --w packages/server"]
+  "packages/**/*.+(js|ts|json)": ["eslint", "prettier"]
 }
