+++ packages/server/src/app/modules/index.ts
@@ -1,2 +1,2 @@
-export * from './auth/auth.module';
+export * from './users/users.module';
 export * from './roles/roles.module';
