+++ packages/server/src/app/modules/roles/roles.service.ts
@@ -2,7 +2,7 @@ import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectModel } from '@nestjs/sequelize';
 import { Role } from './models';
 import { UserRole } from '../../../core/relationships';
-import { User } from '../auth/models';
+import { User } from '../users/models';
 
 @Injectable()
 export class RolesService {
