+++ packages/server/src/core/guards/jwtAt.guard.ts
@@ -2,7 +2,7 @@ import { AuthGuard } from '@nestjs/passport';
 import { JWT_AT } from '../constants';
 
 /**
- * Guard care se ocupa de verificarea tokenului JWT pentru autentificare.
+ * JWT Access Token validation Guard.
  */
 export class JwtAtGuard extends AuthGuard(JWT_AT) {
   constructor() {
